backend/requirements.txt
Django==2.2.16
django-cors-headers==3.11.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
flake8==8.6.0
Pillow==8.3.1
webcolors==1.12